cambio de vista a componente del formulario
