+++ src/views/AboutView.vue
@@ -1,6 +1,8 @@
 <template>
 
   <h1>hola about</h1>
+
+  
 </template>
 
 <style lang="scss" scoped>
